Attempt to install script correctly
